requirements.txt
asgiref==3.8.1
cli_helpers==2.4.0
click==8.2.0
configobj==5.0.9
Django==5.2.1
django-taggit==6.1.0
pgcli==4.3.0
pgspecial==2.2.1
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
Pygments==2.19.1
setproctitle==1.3.6
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.13.2
tzlocal==5.3.1
wcwidth==0.2.13
